tests/requirements-test.txt
# Testing Framework
pytest==7.4.0
pytest-cov==4.1.0
pytest-xdist==3.3.1
pytest-timeout==2.1.0
pytest-randomly==3.13.0
pytest-repeat==0.9.1
pytest-rerunfailures==12.0
pytest-sugar==0.9.7
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-ordering==0.6.0
pytest-clarity==1.0.1
pytest-selenium==4.0.1
pytest-mock==3.11.1
pytest-env==0.8.2
pytest-benchmark==4.0.0
pytest-profiling==1.7.0
pytest-asyncio==0.21.1
pytest-aiohttp==1.0.4
pytest-docker==1.0.1

# Code Coverage
coverage==7.2.7
codecov==2.1.13

# Test Utilities
factory-boy==3.3.0
faker==19.2.0
freezegun==1.2.2
responses==0.23.1
beautifulsoup4==4.12.2
lxml==4.9.3
webtest==3.0.0

# Browser Testing
selenium==4.10.0
webdriver-manager==3.8.6
playwright==1.35.0

# Performance Testing
locust==2.15.1
psutil==5.9.5

# Security Testing
safety==2.3.5
bandit==1.7.5
owasp-zap-api-python==0.4.0

# Database Testing
pytest-postgresql==4.1.1
pytest-mongodb==2.2.0
pytest-redis==3.0.0

# HTTP Testing
requests==2.31.0
httpx==0.24.1
aiohttp==3.8.5

# Mock Services
responses==0.23.1
moto==4.1.11
vcrpy==4.2.1

# Code Quality
black==23.3.0
flake8==6.0.0
mypy==1.4.1
pylint==2.17.4
isort==5.12.0

# Documentation
sphinx==7.0.1
sphinx-rtd-theme==1.2.2
doc8==1.1.1
rstcheck==6.1.1

# Type Checking
types-requests==2.31.0.1
types-PyYAML==6.0.12.11
types-python-dateutil==2.8.19.14
types-beautifulsoup4==4.12.0.6
types-Markdown==3.4.2.10
types-redis==4.6.0.3
types-freezegun==1.1.10
types-aiohttp==3.8.5

# Development Tools
ipdb==0.13.13
ipython==8.14.0
pytest-watch==4.2.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2

# Data Processing
pandas==2.0.3
numpy==1.25.1

# Image Processing
Pillow==10.0.0

# XML Processing
xmltodict==0.13.0
defusedxml==0.7.1

# YAML Processing
PyYAML==6.0.1

# JSON Processing
jsonschema==4.18.4

# Date/Time Utilities
python-dateutil==2.8.2
pytz==2023.3

# File Processing
python-magic==0.4.27
chardet==5.1.0

# Networking
urllib3==2.0.3
websockets==11.0.3

# Cryptography
cryptography==41.0.2
bcrypt==4.0.1

# Compression
zipp==3.15.0

# Progress Bars
tqdm==4.65.0

# Environment Variables
python-dotenv==1.0.0

# CLI Tools
click==8.1.4
colorama==0.4.6

# Debugging
debugpy==1.6.7

# Profiling
line-profiler==4.0.3
memory-profiler==0.61.0

# Dependency Management
pip-tools==7.1.0
pipdeptree==2.9.3

# Version Requirements
# Python >= 3.8
# pip >= 21.0
